Changed cantidad column of ingredientePosicion to decimal

+++ database/migrations/2021_01_20_092941_create_ingrediente_posicions_table.php
@@ -16,7 +16,7 @@ class CreateIngredientePosicionsTable extends Migration
         Schema::create('ingredientePosicion', function (Blueprint $table) {
             $table->integer('posicion')->primary();
             $table->integer('idIngrediente');
-            $table->integer('cantidad');
+            $table->decimal('cantidad', 8, 2);
             $table->timestamps();
         });
     }
